Separate `SyncConfig` injections from production and tests
